requirements.txt
asgiref==3.8.1
contourpy==1.3.0
cycler==0.12.1
diff-match-patch==20241021
Django==4.2.17
django-debug-toolbar==4.4.6
django-import-export==4.3.4
fonttools==4.55.3
gunicorn==23.0.0
importlib_resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
sqlparse==0.5.3
supervisor==4.2.5
tablib==3.7.0
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
zipp==3.21.0
